Add much more info to whitelist.rst
Added the info from Minecraft Wiki as well as correct formatting of the whitelist file itself. This makes it easier for those who need the translated versions of the Sponge Docs. The referenced website doesn't exist in all languages, and those who are in need of a translated documentation will most likely not have much use of a linked website in English.

While manual editing of whitelist.json is generally not desired, it is sometimes necessary. As such, having a ready example of the correct syntax for doing so can prevent much grief for fledgling and veteran server owners alike. I have therefore made an extensive section about it.

I tried my hand at ReStructuredText, but I'm not sure if it will look the same in Sphinx as it does in the online Github editor. If not, I would be grateful if someone with more experience than myself were to take a look and fix it!