Add the Minecraft Wiki information
Continuing my crusade to make things as easy as possible for those who
need the docs translated (albeit with no heed to those poor souls who
have to do the actual translating), I have added the information about
each property found on Minecraft Wiki to this doc page (As a side note,
tables are horrible to work on in Sphinx). More info in PR comment.