[WIP] Condense contributing section
This PR cleans up the mess in /contributing/ leaving just a few pages
in.

This is achieved by removing old (empty) files, merging SpongeForge and
Vanilla into a single implementation page and replacing the Docs files
with a single page.

Comments on this are appreciated. I'll probably create another PR if
there are other sections that need revision too.