Added variable copy for scope to make compilable example
Could not get example to compile as args is not in scope in previous
example; addition of a copy variable full scope results in a functioning
error message return.  Please feel free to discard this and replace with
a better, sponge-uniform superior coding solution, but the example was
indeed nonfunctional before.