Correction to example modification for valid compiling
Example is rewritten in order to compile - args was not in scope to be
able to throw the error return, a copy variable is declared at class
level and set within the constructor so that it may be used to throw the
error.  My first commit of this change was incorrect.