Scheduler updated for new Builder system and methods
Updated scheduler to reflect the newly refactored package location, and
Task.Builder object

Modified Delay and Interval properties to show BOTH methods and replace
no longer accurate "..optional TimeUnits, or else defaults to Ticks..".

Removed information about async use and defaults in absence of
parameters, because it was not clear, and I dont know what of it is
actually correct in the new system.  Someone please re-add asynch
comments for delay/interval descriptions if they are applicable.