Add SpongeDocs License
This is supposed to resolve Issue #565.
Re-licensing the Sponge Documentation requires every significant contributor to accept the new licensing arrangement, by either replying to this thread in the template form, or via email or PM to one of the Editorial staff. We have prepared a table of contributors we need consent from, and that will be added to this PR shortly. It can't be merged until full consent is attained anyway.

Here is the consent form template: 
> I, [developer name] [GitHub email] agree to the licensing of my existing and future contributions to SpongeDocs under the Creative Commons Share-Alike License (CC-BY-SA).

We Apologise for the Inconvenience.