Add warning and workaround for IDEA-194641
IDEA-194641 prevents plugin dependencies from being added to the
classpath when running a plugin from IntelliJ. This results in a
confusing ClassNotFoundException at runtime, which no amount of fiddling
with module dependency settings will resolve.

This commit adds instructions for working around the bug, which involves
creating an entirely new 'container' module in the IntelliJ project.