Change guideline relating to line count for PRs (#798)
* Change guideline relating to line count for PRs

The rule against PRs that are fewer than 20 lines assumes that those PRs are for things like typos or insignificant things. In some cases, as with my only PR to Sponge, it takes 1 line to fix a crash after the author of the offending code has gone to bed. The number of lines is arbitrary, the importance of the change is not. This change better reflects the intent of the original policy.

* Simplify a little of the language

* Update guidelines.rst

Fix line length

* Add link to Minor Issues on GitHub