Added section for merging asset nodes
Created an additional section in the AssetAPI docs for how to merge nodes from an existing Asset to a current configuration node. Have had several people ask for something like this and it seems to be better to set this here than in the configuration node section as chronologically the user may not have looked at assets yet which is required here.