Removed redundant code + added explanations
Previous load from default config was using the AssetAPI but still declared it as a separate URL and had a check to see if the map was empty in an if statement which is a bit more boilerplate than needed. Replaced that with the `copyToFile(path, false, true)` method so its easier to use. Also added explanations as to why you'd use mergeValuesFrom instead of simply copying them as the two have different uses.