Added section for merging asset nodes (#801)
* Added section for merging asset nodes 

Created an additional section in the AssetAPI docs for how to merge nodes from an existing Asset to a current configuration node. Have had several people ask for something like this and it seems to be better to set this here than in the configuration node section as chronologically the user may not have looked at assets yet which is required here.

* Removed previous commit section

will be adding it to under the default config area instead

* Moved merge node section to loaders from assets

Was recommended that this section be moved here for better categorization from my previous commit. Also added an extra note explaining how the nodes will not replace existing nodes but rather only set them up if they are virtual.

* Resolved merge conflict with title being too short

* Re: Fixed merge conflict

Apparently the viewer doesn't reflect how long the underline actually is so its been shorter the entire time

* Re:Re fixed merge conflict

I STG you should never try to fix things when you just wake up. This should fix it up now

* Removed redundant code + added explanations

Previous load from default config was using the AssetAPI but still declared it as a separate URL and had a check to see if the map was empty in an if statement which is a bit more boilerplate than needed. Replaced that with the `copyToFile(path, false, true)` method so its easier to use. Also added explanations as to why you'd use mergeValuesFrom instead of simply copying them as the two have different uses.

* Asset names matching change

Changed the names for the assets to all be default.conf

* Changed headline for syntax + added "Example"

* Wording changes + markup modifications

Removed any excessive headline markup and changed up the wording for why you should use copyToFile() instead of checking for an empty node.

* Changed all DualSpiral Suggestions

Removed redundant line 
Changed things off of example to make entire sections explaining things
Specified that the asset must lie in resources/assets/myplugin as some have been confused where to put it before
Added link to JavaDocs for Asset#copyToFile()

* Fixed syntax